refactor: :recycle: code refactoring
