+++ src/pages/data/show/[responseId].tsx
@@ -5,6 +5,7 @@ import { axGroupList } from "@services/app.service";
 import {
   axGetTemplateResponseById,
   axGetTemplateResponseList,
+  getLoactionInfo,
   getTemplateByShortOrParentName
 } from "@services/cca.service";
 import { getUserIBPsByIds } from "@services/user.service";
@@ -42,13 +43,20 @@ export const getServerSideProps = async (ctx) => {
     data: [header]
   } = await axGetTemplateResponseList({ id: ctx.query.responseId, language: ctx.locale });
 
-  if (response) {
-    header.values.push({
-      type: FORM_TYPE.TEXT,
-      fieldId: "loc",
-      name: "Location",
-      value: capitalize(`${response.location.district}, ${response.location.state}`)
-    });
+  if (header?.centroid?.length) {
+    try {
+      const { data: li, success } = await getLoactionInfo(header.centroid);
+      if (success) {
+        header.values.push({
+          type: FORM_TYPE.TEXT,
+          fieldId: "loc",
+          name: "Location",
+          value: capitalize(`${li.district}, ${li.state}`)
+        });
+      }
+    } catch (e) {
+      console.error(e);
+    }
   }
 
   const canEdit = canEditData([response.userId, ...response.allowedUsers], ctx);
